# Make the dotted substep longer by drawing twice