--- a/resources/charts/observable-plot.html
+++ b/resources/charts/observable-plot.html
@@ -28,6 +28,10 @@
                 <span id="airport-group-size"></span>
                 <input type="range" id="airport-group-size-input" min="0" max="20" value="6" />
             </label>
+            <label>
+                <input type="checkbox" id="dotted-graph-axis-type" />
+                Use a 5th root scale
+            </label>
         </div>
         <div id="chart"></div>
         <script type="module" src="/observable-plot.js"></script>