… resources/charts/dist/assets/plot-27e82af1.js
… resources/charts/dist/observable-plot.html