Camillo asked about this in https://github.com/WebKit/Speedometer/pull/220#issuecomment-1589979335

Now the "dotted" graph is done twice, with 2 different settings. This makes it go to the ballpark of 30ms instead of 18ms on my computer. I tested both Chrome and Firefox, the increase looks similar to me.

[deploy preview](https://make-dotted-longer--speedometer-preview.netlify.app/?suite=Charts-Observable-plot)

What do you think @camillobruni @bgrins @flashdesignory @rniwa ?