@rniwa 2023-06-23
Here are measurements on 16" M2 MacBookPro. I agree this change seems neutral across browsers (the score slightly goes down in all browsers as expected). On the other hand, I'm not really seeing the noise level going down with this change either so there might be not much benefit in taking this change either.

As a general point, there are benefits for different workloads in the benchmark to have varying runtime lengths. If all tests took roughly of a similar magnitude, it can end up favoring a specific runtime behavior such as specific GC timing / strategy and memory allocation pattern / strategy among other things.

* Before: https://speedometer-preview.netlify.app/?suite=Charts-Observable-plot#summary
* After : https://make-dotted-longer--speedometer-preview.netlify.app/?suite=Charts-Observable-plot#home

**Chrome**
* Before
    * 23.3 ± 2.0
    * 23.6 ± 1.9
    * 25.8 ± 2.4
    * 25.7 ± 2.3
    * 26.1 ± 2.2

* After
    * 22.5 ± 2.0
    * 21.9 ± 1.8
    * 20.5 ± 1.5
    * 21.3 ± 1.8
    * 21.0 ± 1.4

**Firefox**
* Before
    * 19.2 ± 1.3
    * 19.7 ± 0.82
    * 19.4 ± 0.97
    * 19.5 ± 0.84
    * 21.6 ± 1.9

* After
    * 17.2 ± 1.1
    * 17.4 ± 0.93
    * 18.8 ± 0.90
    * 17.3 ± 1.1
    * 17.1 ± 1.1

**Safari**
* Before
    * 26.8 ± 2.4
    * 24.5 ± 1.6
    * 24.4 ± 1.7
    * 24.2 ± 1.8
    * 26.2 ± 2.3

* After
    * 21.6 ± 1.8
    * 21.3 ± 2.0
    * 20.9 ± 1.1
    * 23.1 ± 1.9
    * 21.6 ± 1.4

[make-dotted-longer.zip](https://github.com/WebKit/Speedometer/files/11853763/make-dotted-longer.zip)