@bgrins 2023-06-26
I think if it's not reducing noise then I think we should leave it as-is per @rniwa's comment, but would like to hear from @camillobruni who made the original request.