@camillobruni 2023-07-06
I've not don 50 total runs each.

Short step (current behavior):
```
browser                       Google Chrome          Safari                   Firefox
version                       114.0.5735.198         16.5.1 (18615.2.9.11.7)  114.0.2
os                            macos 13.4.1 arm64     macos 13.4.1 arm64       macos 13.4.1 arm64
device                        MacBookPro18,2         MacBookPro18,2           MacBookPro18,2
cpu                           Apple M1 Max 10 cores  Apple M1 Max 10 cores    Apple M1 Max 10 cores
Charts-observable-plot/total  58.6 ± 2.1%            61.3 ± 1.7%              68.07 ± 0.97%
```

Long step (PR):
```
browser                       Google Chrome          Safari                   Firefox
version                       114.0.5735.198         16.5.1 (18615.2.9.11.7)  114.0.2
os                            macos 13.4.1 arm64     macos 13.4.1 arm64       macos 13.4.1 arm64
device                        MacBookPro18,2         MacBookPro18,2           MacBookPro18,2
cpu                           Apple M1 Max 10 cores  Apple M1 Max 10 cores    Apple M1 Max 10 cores
Charts-observable-plot/total  58.36 ± 1.3%           60.99 ± 1.5%             68.20 ± 1.1%
Score                         20.57 ± 1.2%           19.68 ± 1.5%             17.60 ± 1.1%
```