@camillobruni 2023-07-06
I see a minimal improvement on Safari and some more on Chrome, but basically no improvement on Firefox.
Keeping the original 18ms seems fine with me then.